refactor: change double to float in most instances. Removes conversion warnings, and we dont need that level of precision for this project

+++ TruckSim/Core/Utilities.h
@@ -5,12 +5,12 @@
 
 namespace Utilities
 {
-    inline double GetTravelTime(Position a, Position b, float speed)
+    inline float GetTravelTime(Position a, Position b, float speed)
     {
         // Position is in km, speed is in km per hour
         float dx = b.x - a.x;
         float dy = b.y - a.y;
-        double distance = std::sqrt(dx * dx + dy * dy);
+        float distance = std::sqrt(dx * dx + dy * dy);
         return (distance / speed) * 60; // return time in km/minute to fit in time step
     }
 };
